+++ locals.tf
@@ -20,8 +20,8 @@ locals {
   diag_settings_name = coalesce(var.diag_settings_custom_name, "${local.default_name}-diag")
 
   default_app_settings = {
-    "APPLICATION_INSIGHTS_IKEY"      = var.app_insights_instrumentation_key == null ? azurerm_application_insights.app_insights.0.instrumentation_key : var.app_insights_instrumentation_key
-    "APPINSIGHTS_INSTRUMENTATIONKEY" = var.app_insights_instrumentation_key == null ? azurerm_application_insights.app_insights.0.instrumentation_key : var.app_insights_instrumentation_key
+    "APPLICATION_INSIGHTS_IKEY"      = var.app_insights_instrumentation_key == null && var.app_insights_enabled ? azurerm_application_insights.app_insights.0.instrumentation_key : var.app_insights_instrumentation_key
+    "APPINSIGHTS_INSTRUMENTATIONKEY" = var.app_insights_instrumentation_key == null && var.app_insights_enabled ? azurerm_application_insights.app_insights.0.instrumentation_key : var.app_insights_instrumentation_key
   }
 
   app_service_plan_name = split("/", var.app_service_plan_id)[8]